Split compose/network test in unit/integration
Integration tests are ran in the engine instead of mockerode.
